Moved CI tests to test folder
Moved CI test content.
Moved hooks into separate file to avoid running twice from require calls.
Added filter in hook to avoid running for unit tests.
Fixed issue with cloudant init when no backend URL set.
